mlops-week1/pyproject.toml
[tool.poetry]
name = "mlops-week1"
version = "0.1.0"
description = "Week 1: High-performance Image Classification Microservice"
authors = ["Gokul <gokul@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
pydantic = {extras = ["email"], version = "^2.6.3"}
pydantic-settings = "^2.2.1"
torch = "2.2.1"
torchvision = "0.17.1"
python-multipart = "^0.0.9"
pillow = "^10.2.0"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
pytest-asyncio = "^0.23.5"
ruff = "^0.3.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
